triton is up and running with no errors, I upgraded the base image from l4t-ml:r32.6.1-py3 to l4t-ml:r32.7.1-py3

+++ triton/preprocess/1/model.py
@@ -40,7 +40,7 @@ class TritonPythonModel:
                 self.new_shape,
                 auto=False
             )
-            img = img.transpose((2, 0, 1)).astype('float32')
+            img = img.transpose((2, 0, 1)).astype('float16')
             img /= 255
 
             responses.append(
